refactor: verify folder
